Customized the controls of the video player
Added Sound and Speed icons as png to make video player look pleasing to the eyes.

Changed the volume bar's size and changed the playbackrate into a select and option tag so it's easier for users to change the playbackrate to normal.

The controls now show up when video is paused too.